Add reboot type AutomaticRebootSkipStatusReport for update modules
Changelog: Allow update modules to skip reporting status to the server
before rebooting

Before this change, after installing an update module, the mender client
would unconditionally try to notify the server of the client's upcoming
state transition from ArtifactInstall -> ArtifactReboot. This poses a
problem for update modules which (for some reason or another) are expected
to break upstream connectivity for the target device. In this case, the mender
client will "hang" for a while trying to report status until eventually
the client errors out and the update fails.

This change allows an update module to respond with
"AutomaticSkipStatusReport" in order to reboot the host device after
installing the update module while bypassing mender's normal routine
of notifying the server before the state transition into ArtifactReboot.
This option should be used by update modules which have an install stage
that is expected to break upstream connectivity to the mender server.

Signed-off-by: Anthony Green <agreen@starry.com>